refactor: reuse layout
